replace_ranges, scale_thresh: move to util
